+++ pages/auth/registration/index.module.scss
@@ -1,101 +1,4 @@
-@import url('https://fonts.googleapis.com/css2?family=Inter:wght@100;200;300;400;500;600;700;800;900&display=swap'); /*  ?  */
-
 .content {
   background-color: #0d0d0d;
   padding: 24px;
-  .modalContent {
-    color: white;
-  }
-  .registration {
-    max-width: 378px;
-    background-color: #171717;
-    border: 1px solid #333;
-    margin: 0 auto;
-    .modal {
-      position: fixed;
-      top: 0;
-      right: 0;
-      bottom: 0;
-      left: 0;
-      background-color: #00000040;
-      .img {
-        position: fixed;
-        top: 0;
-        right: 0;
-        bottom: 0;
-        left: 0;
-        display: block;
-        margin: auto;
-        height: 40px;
-        width: 40px;
-      }
-    }
-    h1 {
-      color: #fff;
-      text-align: center;
-      font-family: Inter;
-      font-size: 20px;
-      margin-top: 28px;
-    }
-    .item {
-      display: flex;
-      justify-content: center;
-      margin-top: 19px;
-    }
-    .link {
-      padding: 0px 25px;
-    }
-    p {
-      color: #fff;
-      text-align: center;
-      font-family: Inter;
-      margin-top: 20px;
-    }
-    .SignIn {
-      color: #397df6;
-      text-align: center;
-      font-family: Inter;
-      font-weight: 600;
-      display: block;
-      margin: 13px 0 35px;
-    }
-    form {
-      padding: 0px 24px;
-      div {
-        margin-top: 22px;
-      }
-      label {
-        color: #8d9094;
-        font-size: 14px;
-        font-family: Inter;
-        line-height: 24px;
-      }
-      input {
-        box-sizing: border-box; /*  ?  */
-        border-radius: 2px;
-        border: 1px solid #4c4c4c;
-        background-color: #171717;
-        width: 100%;
-        color: #8d9094;
-        font-size: 16px;
-        font-family: Inter;
-        padding: 8px 11px;
-      }
-      .error {
-        border: 1px solid red;
-      }
-      .errorText {
-        color: red;
-        margin: 4px 0px 0px 5px;
-        text-align: left;
-        font-size: 13px;
-      }
-      .button {
-        border-radius: 2px;
-        background-color: #397df6;
-        color: #fff;
-        margin-top: 14px;
-      }
-    }
-  }
 }
